+++ packages/investor-foxy/CHANGELOG.md
@@ -1,3 +1,17 @@
+# [@shapeshiftoss/investor-foxy-v1.1.0](https://github.com/shapeshift/lib/compare/@shapeshiftoss/investor-foxy-v1.0.3...@shapeshiftoss/investor-foxy-v1.1.0) (2022-03-25)
+
+
+### Features
+
+* foxy rebase history ([#492](https://github.com/shapeshift/lib/issues/492)) ([11b8270](https://github.com/shapeshift/lib/commit/11b82703573799fb9575029c6c2d4d40269213d3))
+
+# [@shapeshiftoss/investor-foxy-v1.0.3](https://github.com/shapeshift/lib/compare/@shapeshiftoss/investor-foxy-v1.0.2...@shapeshiftoss/investor-foxy-v1.0.3) (2022-03-25)
+
+
+### Bug Fixes
+
+* don't use strings in contract calls ([#489](https://github.com/shapeshift/lib/issues/489)) ([8b7626c](https://github.com/shapeshift/lib/commit/8b7626c92a65f4821156f24412e1db9c7919720f))
+
 # [@shapeshiftoss/investor-foxy-v1.0.2](https://github.com/shapeshift/lib/compare/@shapeshiftoss/investor-foxy-v1.0.1...@shapeshiftoss/investor-foxy-v1.0.2) (2022-03-23)
 
 # [@shapeshiftoss/investor-foxy-v1.0.1](https://github.com/shapeshift/lib/compare/@shapeshiftoss/investor-foxy-v1.0.0...@shapeshiftoss/investor-foxy-v1.0.1) (2022-03-22)